Validaciones

+++ src/comments/services/comments.service.ts
@@ -23,10 +23,10 @@ export class CommentsService {
         const student2 = await this.studentrepo.findOne({where: {id_Student: id_Student}});
         console.log(lesson)
         console.log(student2)
-        if ((!lesson) || (!student2)) {
-            throw new Notification ('Leccion no encontrada')
+        if ((!lesson) || (!student2)){
+            return ('Leccion no encontrada')
         }
-        if (student2.state='Active'){
+        if (student2.state=='Active'){
             const newComment = new Comments();
             //newComment.id = data.id;
             newComment.text = data.text;
@@ -34,13 +34,14 @@ export class CommentsService {
             newComment.lesson = lesson;
             return this.commentrepo.save(newComment);
         }
-        else return;
+        else return("No puede comentar, estas bloqueado");
    }
 
     async delete(id_s:number, id_c:number){ 
         console.log('estudiante: '+ id_s)
         console.log('leccion ' + id_c)
         const qb = await this.commentrepo.createQueryBuilder("comments").where("comments.lesson_id = :id_c and comments.student_id =:id_s").setParameters({id_c,id_s}).getOne();
+        if (!qb){ return ("No se encontro el comentario")}
         await this.commentrepo.delete(qb.id);
         return ('Comentario Eliminado')
     }
